42 Extracted config class
