+++ jobs/src/test/java/com/marklogic/spring/batch/config/ImportShapefilesTest.java
@@ -2,6 +2,7 @@ package com.marklogic.spring.batch.config;
 
 import com.marklogic.client.helper.LoggingObject;
 import com.marklogic.spring.batch.item.DirectoryReader;
+import com.marklogic.spring.batch.item.shapefile.ImportShapefilesConfig;
 import com.marklogic.spring.batch.item.shapefile.ShapefileAndJson;
 import com.marklogic.spring.batch.item.shapefile.ShapefileAndJsonWriter;
 import com.marklogic.spring.batch.item.shapefile.ShapefileProcessor;
@@ -28,29 +29,8 @@ public class ImportShapefilesTest extends AbstractJobsJobTest {
 
     @Test
     public void test() {
-        runJob(ImportShapefilesConfig.class, "--input_file_path", "build/shapefiles");
+        runJob(ImportShapefilesConfig.class,
+                "--input_file_path", "build/shapefiles");
     }
 
-    public static class ImportShapefilesConfig extends AbstractMarkLogicBatchConfig {
-
-        @Bean
-        public Job job(@Qualifier("step1") Step step1) {
-            return jobBuilderFactory.get("importShapefilesJob").start(step1).build();
-        }
-
-        @Bean
-        @JobScope
-        protected Step step1(
-                @Value("#{jobParameters['input_file_path']}") String inputFilePath) {
-
-            DirectoryReader reader = new DirectoryReader(new File(inputFilePath));
-
-            return stepBuilderFactory.get("step1")
-                    .<File, ShapefileAndJson>chunk(10)
-                    .reader(reader)
-                    .processor(new ShapefileProcessor())
-                    .writer(new ShapefileAndJsonWriter(getDatabaseClient()))
-                    .build();
-        }
-    }
 }
